requirements.txt
Django==4.2.3
djangorestframework==3.14.0
django-filter==23.2
djoser==2.2.0
Pillow==10.0.0
coreapi==2.3.3
pytest==7.4.0
pytest-django==4.5.2
pytest-drf==1.1.3




